+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Builtin Prosody: optional settings to change the url for Prosody's API calls.
 * Diagnostic tool: testing API communication from Peertube to Prosody.
 * Diagnostic tool: testing API communication from Prosody to Peertube.
+* Diagnostic tool: correctly parse Prosody nightly build versions.
 
 ## v3.0.0